+++ app/src/main/java/com/example/roomfit/presentation/mate/UserInfo.kt
@@ -1,7 +1,6 @@
 package com.example.roomfit.presentation.mate
 
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -14,17 +13,13 @@ import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.example.roomfit.R
 import com.example.roomfit.ui.theme.ComponentBeige
-import com.example.roomfit.ui.theme.LoginButton
 import com.example.roomfit.ui.theme.UserInfo
-import com.example.roomfit.ui.theme.bodyWriting
 
 @Preview
 @Composable
